+++ asn1/AbstractKey.c
@@ -5,35 +5,35 @@
  * 	`asn1c -fwide-types -findirect-choice`
  */
 
-#include "KeyResource.h"
+#include "AbstractKey.h"
 
-static asn_TYPE_member_t asn_MBR_KeyResource_1[] = {
-    { ATF_NOFLAGS, 0, myoffsetof(struct KeyResource, choice.counter),
-        (ASN_TAG_CLASS_UNIVERSAL | (2 << 2)),
+static asn_TYPE_member_t asn_MBR_AbstractKey_1[] = {
+    { ATF_POINTER, 0, myoffsetof(struct AbstractKey, choice.secretKey),
+        (ASN_TAG_CLASS_UNIVERSAL | (16 << 2)),
         0,
-        &asn_DEF_CipherKeyResource,
+        &asn_DEF_SecretKeyData,
         0,	/* Defer constraints checking to the member type */
         0,	/* PER is not compiled, use -gen-PER */
         0,
-        "counter"
+        "secretKey"
         },
 };
-static const asn_TYPE_tag2member_t asn_MAP_KeyResource_tag2el_1[] = {
-    { (ASN_TAG_CLASS_UNIVERSAL | (2 << 2)), 0, 0, 0 } /* counter */
+static const asn_TYPE_tag2member_t asn_MAP_AbstractKey_tag2el_1[] = {
+    { (ASN_TAG_CLASS_UNIVERSAL | (16 << 2)), 0, 0, 0 } /* secretKey */
 };
-static asn_CHOICE_specifics_t asn_SPC_KeyResource_specs_1 = {
-    sizeof(struct KeyResource),
-    offsetof(struct KeyResource, _asn_ctx),
-    offsetof(struct KeyResource, present),
-    sizeof(((struct KeyResource *)0)->present),
-    asn_MAP_KeyResource_tag2el_1,
+static asn_CHOICE_specifics_t asn_SPC_AbstractKey_specs_1 = {
+    sizeof(struct AbstractKey),
+    offsetof(struct AbstractKey, _asn_ctx),
+    offsetof(struct AbstractKey, present),
+    sizeof(((struct AbstractKey *)0)->present),
+    asn_MAP_AbstractKey_tag2el_1,
     1,	/* Count of tags in the map */
     0,
     -1	/* Extensions start */
 };
-asn_TYPE_descriptor_t asn_DEF_KeyResource = {
-    "KeyResource",
-    "KeyResource",
+asn_TYPE_descriptor_t asn_DEF_AbstractKey = {
+    "AbstractKey",
+    "AbstractKey",
     CHOICE_free,
     CHOICE_print,
     CHOICE_constraint,
@@ -48,8 +48,8 @@ asn_TYPE_descriptor_t asn_DEF_KeyResource = {
     0,	/* No tags (pointer) */
     0,	/* No tags (count) */
     0,	/* No PER visible constraints */
-    asn_MBR_KeyResource_1,
+    asn_MBR_AbstractKey_1,
     1,	/* Elements count */
-    &asn_SPC_KeyResource_specs_1	/* Additional specs */
+    &asn_SPC_AbstractKey_specs_1	/* Additional specs */
 };