+++ source/ak_tools.h
@@ -34,6 +34,7 @@
 
 /* ----------------------------------------------------------------------------------------------- */
  #include <libakrypt.h>
+ #include <SecretKeyData.h>
 
 /* ----------------------------------------------------------------------------------------------- */
 /*! \brief Неэкспортируемая функция установления уровня аудита. */
@@ -43,6 +44,14 @@
  int ak_libakrypt_set_option( const char *name, const int value );
 /*! \brief Функция возвращает значение опции с заданным именем. */
  int ak_libakrypt_get_option( const char *name );
+
+/* ----------------------------------------------------------------------------------------------- */
+/*! \brief Функция выделяет память для указателя OCTET_STRING_t и копирует в нее данные, хранящиеся в ptr */
+ int ak_ptr_to_asn1_octet_string( const ak_pointer , const size_t , OCTET_STRING_t * );
+/*! \brief Очистка структуры `SecretKeyData` после вызова функции ak_skey_to_asn1_secret_key_data() */
+ int ak_asn1_secret_key_data_destroy( SecretKeyData_t * );
+/*! \brief Сохранение asn1 структуры в файле в der-представлении */
+ int ak_asn1_save_to_der_file( struct asn_TYPE_descriptor_s *, void *, const char * );
 
  #endif
 /* ----------------------------------------------------------------------------------------------- */